requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
chime==0.7.0
dateparser==1.1.6
Django==4.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
ntplib==0.4.0
numpy==1.25.0
pandas==2.0.2
pycryptodome==3.18.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
regex==2022.10.31
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
TA-Lib @ file:///E:/MY%20PYTHON%20PROJECTS/BOT_SERVER_AND_SIGNAL/BOT_SIGNAL_GENERATOR/TA_Lib-0.4.24-cp310-cp310-win_amd64.whl
tradingview-ta==3.3.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.3
websocket-client==1.6.1
websockets==11.0.3
yarl==1.9.2
